+++ README.md
@@ -10,19 +10,23 @@ Learning rate free learning for SGD, AdaGrad and Adam!
 
 The provided Pytorch Optimizer classes are drop-in replacements, either copy into your project or use via pip with dadaptation.DAdaptSGD,  dadaptation.DAdaptAdam or dadaptation.DAdaptAdaGrad.
 
- - **Set the LR parameter to 1.0**. This parameter is not ignored, rather, setting it larger to smaller will directly scale up or down the D-Adapted learning rate. 
+ - **Set the LR parameter to 1.0**. This parameter is not ignored. Setting it larger to smaller will directly scale up or down the D-Adapted learning rate estimate.
  - **Use the same learning rate scheduler you would normally use on the problem.**
  - The Adam variant supports AdamW style weight decay, just set decouple=True. It is not turned on by default, so if you are replacing your adam implementation, make sure you use decoupled if necessary.
  - It may be necessary to use larger weight decay than you would normally use, try a factor of 2 or 4 bigger if you see overfitting. D-Adaptation uses larger learning rates than people typically hand-choose, in some cases that requires more decay.
  - Use the log_every setting to see the learning rate being used (d*lr) and the current D bound.
  - Only the AdaGrad version supports sparse gradients.
- - The Adam IP variant implements a tighter D bound, which may help on some problems. The IP variants should be considered experimental.
  - Parameter-group level LR values are not fully supported. The optimizer only supports setting zero LR for some groups in order to do fine-tuning on parts of a model.
  
 ## Change Log
 
+### Version 3.0
+ - Major improvements to DAdaptAdam, improving the performance particularly on Transformer models. This variant may behave differently in practice. The old version is availiable under experimental/d_adapt_adam_preprint if you wish to continue to use it.
+ - The IP variant is now the main variant of the method.
+ - Added Lion. This is highly experimental. Feedback on it's performance is welcome.
+
 ### Version 2.0
- - Added DAdaptAdan - should still be considered experimental.
+ - Added Adan - should still be considered experimental.
  - Added support for PyTorch's Fully Sharded Data Parallel. 
  - Improved support of edge cases such as learning rate zero.
  - Improved logging - uses Python logging rather than print statements
