+++ dadaptation/dadapt_adam.py
@@ -19,15 +19,15 @@ if TYPE_CHECKING:
 else:
     _params_t = Any
 
-def to_real(x):
-    if torch.is_complex(x):
-        return x.real
-    else:
-        return x
-
-class DAdaptAdam(torch.optim.Optimizer):
+class DAdaptAdamIP(torch.optim.Optimizer):
     r"""
-    Implements Adam with D-Adaptation automatic step-sizes. Leave LR set to 1 unless you encounter instability.
+    Implements Adam with D-Adaptation automatic step-sizes. 
+    Leave LR set to 1 unless you encounter instability.
+
+    This IP variant uses a tighter bound than the non-IP version,
+    and so will typically choose larger step sizes. It has not 
+    been as extensively tested.
+    
     Arguments:
         params (iterable): 
             Iterable of parameters to optimize or dicts defining parameter groups.
@@ -36,7 +36,7 @@ class DAdaptAdam(torch.optim.Optimizer):
         betas (Tuple[float, float], optional): coefficients used for computing
             running averages of gradient and its square (default: (0.9, 0.999))
         eps (float): 
-            Term added to the denominator outside of the root operation to improve numerical stability. (default: 1e-8).
+            Term added to the denominator outside of the root operation to improve numerical stability. (default: 0).
         weight_decay (float): 
             Weight decay, i.e. a L2 penalty (default: 0).
         log_every (int): 
@@ -55,13 +55,10 @@ class DAdaptAdam(torch.optim.Optimizer):
             than PyTorch's builtin version, the auto-detection won't work.
     """
     def __init__(self, params, lr=1.0, 
-                 betas=(0.9, 0.999), 
-                 eps=1e-8,
-                 weight_decay=0, 
-                 log_every=0,
+                 betas=(0.9, 0.999), eps=1e-8,
+                 weight_decay=0, log_every=0,
                  decouple=False,
-                 d0=1e-6, 
-                 growth_rate=float('inf'),
+                 d0=1e-6, growth_rate=float('inf'),
                  fsdp_in_use=False):
         if not 0.0 < d0:
             raise ValueError("Invalid d0 value: {}".format(d0))
@@ -77,16 +74,17 @@ class DAdaptAdam(torch.optim.Optimizer):
         if decouple:
             print(f"Using decoupled weight decay")
 
+        
         defaults = dict(lr=lr, betas=betas, eps=eps,
                         weight_decay=weight_decay,
                         d = d0, 
                         k=0, 
-                        gsq_weighted=0.0,
+                        numerator_weighted=0.0,
                         log_every=log_every,
-                        decouple=decouple,
                         growth_rate=growth_rate,
+                        decouple=decouple,
                         fsdp_in_use=fsdp_in_use)
-        
+        self.d0 = d0
         super().__init__(params, defaults)
 
     @property
@@ -108,40 +106,36 @@ class DAdaptAdam(torch.optim.Optimizer):
         if closure is not None:
             loss = closure()
 
-
-        g_sq = 0.0
-        sksq_weighted = 0.0
         sk_l1 = 0.0
 
+        group = self.param_groups[0]
+        numerator_weighted = group['numerator_weighted']
+        d = group['d']
         lr = max(group['lr'] for group in self.param_groups)
 
-        group = self.param_groups[0]
-        gsq_weighted = group['gsq_weighted']
-        d = group['d']
         dlr = d*lr
         
         growth_rate = group['growth_rate']
         decouple = group['decouple']
-        fsdp_in_use = group['fsdp_in_use']
         log_every = group['log_every']
+        fsdp_in_use = group['fsdp_in_use']
 
         beta1, beta2 = group['betas']
         sqrt_beta2 = beta2**(0.5)
 
+        numerator_acum = 0.0
+
         for group in self.param_groups:
-            group_lr = group['lr']
             decay = group['weight_decay']
             k = group['k']
             eps = group['eps']
-
-            if group_lr not in [lr, 0.0]:
-                raise RuntimeError(f"Setting different lr values in different parameter groups is only supported for values of 0")
 
             for p in group['params']:
                 if p.grad is None:
                     continue
                 if hasattr(p, "_fsdp_flattened"):
                     fsdp_in_use = True
+                
                 grad = p.grad.data
                 
                 # Apply weight decay (coupled variant)
@@ -153,69 +147,59 @@ class DAdaptAdam(torch.optim.Optimizer):
                 # State initialization
                 if 'step' not in state:
                     state['step'] = 0
-                    state['s'] = torch.zeros_like(p.data, memory_format=torch.preserve_format).detach()
+                    state['s'] = torch.zeros_like(p.data).detach()
                     # Exponential moving average of gradient values
-                    state['exp_avg'] = torch.zeros_like(p.data, memory_format=torch.preserve_format).detach()
+                    state['exp_avg'] = torch.zeros_like(p.data).detach()
                     # Exponential moving average of squared gradient values
-                    state['exp_avg_sq'] = torch.zeros_like(to_real(p.data), memory_format=torch.preserve_format).detach()
+                    state['exp_avg_sq'] = torch.zeros_like(p.data).detach()
 
                 exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
                 
-                grad_grad = to_real(grad * grad.conj())
+                s = state['s']
+
+                denom = exp_avg_sq.sqrt().add_(eps)
+                numerator_acum += dlr * torch.dot(grad.flatten(), s.div(denom).flatten()).item()
 
                 # Adam EMA updates
-                if group_lr > 0:
-                    exp_avg.mul_(beta1).add_(grad, alpha=dlr*(1-beta1))
-                    exp_avg_sq.mul_(beta2).add_(grad_grad, alpha=1-beta2)
-                    
-                    denom = exp_avg_sq.sqrt().add_(eps)
+                exp_avg.mul_(beta1).add_(grad, alpha=dlr*(1-beta1))
+                exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1-beta2)
 
-                    g_sq += grad_grad.div_(denom).sum().item()
-
-                    s = state['s']
-                    s.mul_(sqrt_beta2).add_(grad, alpha=dlr*(1-sqrt_beta2))
-                    sksq_weighted += to_real(s * s.conj()).div_(denom).sum().item()
-                    sk_l1 += s.abs().sum().item()
+                s.mul_(sqrt_beta2).add_(grad, alpha=dlr*(1-sqrt_beta2))
+                sk_l1 += s.abs().sum().item()
 
             ######
 
-        gsq_weighted = sqrt_beta2*gsq_weighted + (1-sqrt_beta2)*(dlr**2)*g_sq
+        numerator_weighted = sqrt_beta2*numerator_weighted + (1-sqrt_beta2)*numerator_acum
         d_hat = d
 
         # if we have not done any progres, return
         # if we have any gradients available, will have sk_l1 > 0 (unless \|g\|=0)
         if sk_l1 == 0:
             return loss
-
+        
         if lr > 0.0:
             if fsdp_in_use:
-                dist_tensor = torch.zeros(3).cuda()
-                dist_tensor[0] = sksq_weighted
-                dist_tensor[1] = gsq_weighted
-                dist_tensor[2] = sk_l1
+                dist_tensor = torch.zeros(2).cuda()
+                dist_tensor[0] = numerator_weighted
+                dist_tensor[1] = sk_l1
                 dist.all_reduce(dist_tensor, op=dist.ReduceOp.SUM)
-                global_sksq_weighted = dist_tensor[0]
-                global_gsq_weighted = dist_tensor[1]
-                global_sk_l1 = dist_tensor[2]
+                global_numerator_weighted = dist_tensor[0]
+                global_sk_l1 = dist_tensor[1]
             else:
-                global_sksq_weighted = sksq_weighted
-                global_gsq_weighted = gsq_weighted
+                global_numerator_weighted = numerator_weighted
                 global_sk_l1 = sk_l1
 
-            d_hat = 0.5*(global_sksq_weighted/(1-sqrt_beta2) - global_gsq_weighted)/global_sk_l1
+
+            d_hat = global_numerator_weighted/((1-sqrt_beta2)*global_sk_l1)
             d = max(d, min(d_hat, d*growth_rate))
 
         if log_every > 0 and k % log_every == 0:
-            logging.info(
-                f"(k={k}) dlr: {dlr:1.1e} d_hat: {d_hat:1.1e}, d: {d:1.8}. "
-                f"sksq_weighted={global_sksq_weighted:1.1e} gsq_weighted={global_gsq_weighted:1.1e} "
-                f"sk_l1={global_sk_l1:1.1e}{' (FSDP)' if fsdp_in_use else ''}")
+            logging.info(f"lr: {lr} dlr: {dlr} d_hat: {d_hat}, d: {d}. sk_l1={global_sk_l1:1.1e} numerator_weighted={global_numerator_weighted:1.1e}")
 
         for group in self.param_groups:
-            group['gsq_weighted'] = gsq_weighted
+            group['numerator_weighted'] = numerator_weighted
             group['d'] = d
 
-            group_lr = group['lr']
             decay = group['weight_decay']
             k = group['k']
             eps = group['eps']
@@ -232,10 +216,9 @@ class DAdaptAdam(torch.optim.Optimizer):
                 state['step'] += 1
 
                 denom = exp_avg_sq.sqrt().add_(eps)
-                denom = denom.type(p.type())
 
                 # Apply weight decay (decoupled variant)
-                if decay != 0 and decouple and group_lr > 0:
+                if decay != 0 and decouple:
                     p.data.add_(p.data, alpha=-decay * dlr)
