+++ dadaptation/dadapt_sgd.py
@@ -13,7 +13,8 @@ import torch.distributed as dist
 
 class DAdaptSGD(torch.optim.Optimizer):
     r"""
-    Implements SGD with D-Adaptation automatic step-sizes. Leave LR set to 1 unless you encounter instability.
+    Implements Adam with D-Adaptation automatic step-sizes. Leave LR set to 1 unless you encounter instability.
+
     Arguments:
         params (iterable): 
             Iterable of parameters to optimize or dicts defining parameter groups.
@@ -46,18 +47,16 @@ class DAdaptSGD(torch.optim.Optimizer):
         d0=1e-6, growth_rate=float('inf'),
         fsdp_in_use=False):
 
-        if d0 <= 0:
+        if not 0.0 < d0:
             raise ValueError("Invalid d0 value: {}".format(d0))
-        if lr <= 0:
+        if not 0.0 < lr:
             raise ValueError("Invalid learning rate: {}".format(lr))
 
         defaults = dict(lr=lr,
             momentum=momentum, 
-            weight_decay=weight_decay, 
-            k=0,
+            weight_decay=weight_decay, k=0,
             log_every=log_every,
-            gsq_weighted=0.0, 
-            g0_norm=0.0,
+            numerator_weighted=0.0, 
             d=d0, 
             growth_rate=growth_rate,
             fsdp_in_use=fsdp_in_use)
@@ -68,8 +67,6 @@ class DAdaptSGD(torch.optim.Optimizer):
         except:
             self.rank = 0
 
-        #self.group = dist.new_group(range(dist.get_world_size()))
-
         super().__init__(params, defaults)
 
     def step(self, closure=None):
@@ -77,53 +74,45 @@ class DAdaptSGD(torch.optim.Optimizer):
         if closure is not None:
             loss = closure()
 
+        group = self.param_groups[0]
         lr = max(group['lr'] for group in self.param_groups)
 
-        # Stored at the group level to ensure resuming from checkpoints works
-        group = self.param_groups[0]
         decay = group['weight_decay']
         momentum = group['momentum']
         log_every = group['log_every']
         ck = 1 - momentum
         k = group['k']
 
-        gsq_weighted = group['gsq_weighted']
-        g0_norm = group['g0_norm']
+        numerator_weighted = group['numerator_weighted']
         growth_rate = group['growth_rate']
-        fsdp_in_use = group['fsdp_in_use']
         d = group['d']
-
-        g_sq = 0.0
+        fsdp_in_use = group['fsdp_in_use']
+        
+        group = self.param_groups[0]
+        
         sk_sq = 0.0
 
-        for group in self.param_groups:
-            group_lr = group['lr']
-            if group_lr not in [lr, 0.0]:
-                raise RuntimeError(f"Setting different lr values in different parameter groups is only supported for values of 0")
-            for p in group['params']:
-                if p.grad is None:
-                    continue
-                if hasattr(p, "_fsdp_flattened"):
-                    fsdp_in_use = True
-                grad = p.grad.data
-                
-                # Apply weight decay
-                if decay != 0:
-                    if grad.is_sparse:
-                        raise RuntimeError("weight_decay option is not compatible with sparse gradients")
+        if k == 0: 
+            g_sq = 0.0
+            for group in self.param_groups:
+                for p in group['params']:
+                    if p.grad is None:
+                        continue
+                    if hasattr(p, "_fsdp_flattened"):
+                        fsdp_in_use = True
+                    grad = p.grad.data
+                    
+                    # Apply weight decay
+                    if decay != 0:
+                        if grad.is_sparse:
+                            raise RuntimeError("weight_decay option is not compatible with sparse gradients")
 
-                    grad.add_(p.data, alpha=decay)
+                        grad.add(p.data, alpha=decay)
 
-                state = self.state[p]
+                    state = self.state[p]
 
-                if group_lr > 0.0:
                     g_sq += (grad * grad).sum().item()
 
-        # if we have any gradients available, will have g_sq > 0 (unless \|g\|=0)
-        if g_sq == 0:
-            return loss
-
-        if k == 0: 
             if fsdp_in_use:
                 dist_tensor = torch.zeros(1).cuda()
                 dist_tensor[0] = g_sq
@@ -136,7 +125,6 @@ class DAdaptSGD(torch.optim.Optimizer):
         dlr = d*lr/g0_norm
 
         for group in self.param_groups:
-            group_lr = group['lr']
             for p in group['params']:
                 if p.grad is None:
                     continue
@@ -148,40 +136,49 @@ class DAdaptSGD(torch.optim.Optimizer):
                     s = state['s'] = torch.zeros_like(p.data).detach()
                     x0 = state['x0'] = torch.clone(p.data).detach()
 
+                    # Apply weight decay
+                    if decay != 0:
+                        if grad.is_sparse:
+                            raise RuntimeError("weight_decay option is not compatible with sparse gradients")
+
+                        grad.add_(p.data, alpha=decay)
+
                 s = state['s']
 
-                if group_lr > 0.0:
-                    s.data.add_(grad, alpha=dlr)
+                numerator_weighted += dlr * torch.dot(grad.flatten(), s.flatten()).item()
                 
+                s.data.add_(grad, alpha=dlr)
                 sk_sq += (s * s).sum().item()
             ######
 
-        gsq_weighted = gsq_weighted + dlr*dlr*g_sq
         d_hat = d
 
         if lr > 0.0:
             if fsdp_in_use:
                 dist_tensor = torch.zeros(2).cuda()
                 dist_tensor[0] = sk_sq
-                dist_tensor[1] = gsq_weighted
+                dist_tensor[1] = numerator_weighted
                 dist.all_reduce(dist_tensor, op=dist.ReduceOp.SUM)
                 global_sk_sq = dist_tensor[0]
-                global_gsq_weighted = dist_tensor[1]
+                global_numerator_weighted = dist_tensor[1]
             else:
                 global_sk_sq = sk_sq
-                global_gsq_weighted = gsq_weighted
+                global_numerator_weighted = numerator_weighted
 
-            d_hat = (global_sk_sq - global_gsq_weighted)/(math.sqrt(global_sk_sq))
-            d = group['d'] = max(d, min(d_hat, d*growth_rate))
+            d_hat = 2*global_numerator_weighted/math.sqrt(global_sk_sq)
+            d = max(d, min(d_hat, d*growth_rate))
+
+
+        # if we have not done any updates
+        # if we have any gradients available, will have sk_sq > 0 (unless \|g\|=0)
+        if global_sk_sq == 0:
+            return loss
 
         if log_every > 0 and k % log_every == 0:
-            logging.info(
-                f"(k={k}) dlr: {dlr:1.1e} d_hat: {d_hat:1.1e}, d: {d:1.8}. "
-                f"sk_sq={global_sk_sq:1.1e} gsq_weighted={global_gsq_weighted:1.1e} "
-                f"g0_norm={g0_norm:1.1e}{' (FSDP)' if fsdp_in_use else ''}")
+            logging.info(f"(r={self.rank},k={k}) dlr: {dlr} d_hat: {d_hat}, d: {d}. sk_norm={math.sqrt(global_sk_sq)} numerator_weighted={global_numerator_weighted} g0_norm={g0_norm}", flush=True)
 
         for group in self.param_groups:
-            group['gsq_weighted'] = gsq_weighted
+            group['numerator_weighted'] = numerator_weighted
             group['d'] = d
             group['g0_norm'] = g0_norm
             ######################################
