+++ dadaptation/experimental/__init__.py
@@ -1,3 +1,2 @@
-from .dadapt_adam_ip import DAdaptAdamIP
-from .dadapt_sgd_ip import DAdaptSGDIP
 from .dadapt_adan_ip import DAdaptAdanIP
+from .dadapt_adam_preprint import DAdaptAdamPreprint
